code refactor
