+++ lib/metanorma/standoc/isodoc.rng
@@ -17,7 +17,7 @@
   these elements; we just want one namespace for any child grammars
   of this.
 -->
-<!-- VERSION v2.0.2 -->
+<!-- VERSION v2.0.3 -->
 <grammar xmlns:a="http://relaxng.org/ns/compatibility/annotations/1.0" xmlns="http://relaxng.org/ns/structure/1.0" datatypeLibrary="http://www.w3.org/2001/XMLSchema-datatypes">
   <include href="reqt.rng"/>
   <include href="basicdoc.rng">
@@ -424,19 +424,25 @@ normative or informative references, some split references into sections organiz
       </choice>
     </define>
     <define name="TdAttributes" combine="interleave">
-      <attribute name="style">
-        <a:documentation>CSS style: only background-color supported</a:documentation>
-      </attribute>
+      <optional>
+        <attribute name="style">
+          <a:documentation>CSS style: only background-color supported</a:documentation>
+        </attribute>
+      </optional>
     </define>
     <define name="ThAttributes" combine="interleave">
-      <attribute name="style">
-        <a:documentation>CSS style: only background-color supported</a:documentation>
-      </attribute>
+      <optional>
+        <attribute name="style">
+          <a:documentation>CSS style: only background-color supported</a:documentation>
+        </attribute>
+      </optional>
     </define>
     <define name="TrAttributes">
-      <attribute name="style">
-        <a:documentation>CSS style: only background-color supported</a:documentation>
-      </attribute>
+      <optional>
+        <attribute name="style">
+          <a:documentation>CSS style: only background-color supported</a:documentation>
+        </attribute>
+      </optional>
     </define>
     <define name="table-note">
       <element name="note">
